+++ builds/4chan-X.user.js
@@ -9920,7 +9920,7 @@
       }
     },
     initReady: function() {
-      var board, boardChild, err, errors, href, posts, thread, threadChild, threads, _i, _j, _len, _len1, _ref, _ref1;
+      var board, boardChild, err, errors, href, passLink, posts, styleSelector, thread, threadChild, threads, _i, _j, _len, _len1, _ref, _ref1;
 
       if (d.title === '4chan - 404 Not Found') {
         if (Conf['404 Redirect'] && g.VIEW === 'thread') {
@@ -9975,6 +9975,16 @@
           $.event('4chanXInitFinished');
           return Main.checkUpdate();
         });
+        if (styleSelector = $.id('styleSelector')) {
+          passLink = $.el('a', {
+            textContent: '4chan Pass',
+            href: 'javascript:;'
+          });
+          $.on(passLink, 'click', function() {
+            return window.open('//sys.4chan.org/auth', 'This will steal your data.', 'left=0,top=0,width=500,height=255,toolbar=0,resizable=0');
+          });
+          $.before(styleSelector.previousSibling, [$.tn('['), passLink, $.tn(']\u00A0\u00A0')]);
+        }
         return;
       }
       $.event('4chanXInitFinished');