+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 seaweedchan
 - Fix boards with previously deleted archives not switching to new archives 
 
+ihavenoface:
+- 4chan Pass link by the style selector
+
 ### 1.1.8 - 2013-05-01
 seaweedchan:
 - Fix QR not clearing on submit with Posting Success Notifications disabled